draft for restructuring consts/utils, removing old files
