+++ src/constants/index.js
@@ -1,0 +1,70 @@
+import {
+  CheckCircleIcon,
+  ExclamationCircleIcon,
+  ExclamationTriangleIcon,
+  BellIcon,
+  PauseCircleIcon,
+  CircleNotchIcon,
+  UnknownIcon,
+} from '@patternfly/react-icons';
+
+import dangerColor from '@patternfly/react-tokens/dist/esm/global_danger_color_100';
+import warningColor from '@patternfly/react-tokens/dist/esm/global_warning_color_100';
+import successColor from '@patternfly/react-tokens/dist/esm/global_success_color_100';
+import infoColor from '@patternfly/react-tokens/dist/esm/global_info_color_100';
+
+export const statusMapper = [
+  'done',
+  'error',
+  'pending',
+  'unknown',
+  'updating',
+  'warning',
+  'notification',
+];
+
+export const statusToIcon = {
+  done: {
+    icon: CheckCircleIcon,
+    color: successColor.value,
+    title: 'Fully adopted',
+  },
+  error: {
+    icon: ExclamationCircleIcon,
+    color: dangerColor.value,
+    title: 'Error while adopting',
+  },
+  pending: { icon: PauseCircleIcon, title: 'Pending adoption' },
+  updating: { icon: CircleNotchIcon, title: 'Updating' },
+  unknown: { icon: UnknownIcon, title: 'Unknown state' },
+  warning: {
+    icon: ExclamationTriangleIcon,
+    color: warningColor.value,
+    title: 'Warning while adopting',
+  },
+  notification: {
+    icon: BellIcon,
+    color: infoColor.value,
+    title: 'Delivering',
+  },
+};
+
+export const sortByDirection = (data, direction = 'asc') =>
+  data.sort((a, b) =>
+    direction === 'asc'
+      ? a.name.toLowerCase().localeCompare(b.name.toLowerCase())
+      : b.name.toLowerCase().localeCompare(a.name.toLowerCase())
+  );
+
+export const deviceSummaryMapper = [
+  'active',
+  'noReports',
+  'neverReported',
+  'orphaned',
+];
+
+export const imageDistributionMapper = {
+  'rhel-8.1': 'RHEL 8.1',
+  'rhel-8.2': 'RHEL 8.2',
+  'rhel-8.3': 'RHEL 8.3',
+};
